+++ app/src/main/java/com/oounabaramusic/android/service/MusicPlayService.java
@@ -11,16 +11,15 @@ import android.os.Message;
 import android.preference.PreferenceManager;
 import android.widget.Toast;
 
-import com.oounabaramusic.android.LocalMusicActivity;
 import com.oounabaramusic.android.bean.Music;
 import com.oounabaramusic.android.dao.LocalMusicDao;
-import com.oounabaramusic.android.okhttputil.HttpUtil;
 import com.oounabaramusic.android.util.LogUtil;
-import com.oounabaramusic.android.util.MyEnvironment;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.Date;
+import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Random;
@@ -30,18 +29,20 @@ import androidx.annotation.Nullable;
 public class MusicPlayService extends Service {
 
     public static final int NOT_PREPARE=0;
+    public static final int PREPAREING=5;
     public static final int IS_PREPARE=1;
     public static final int IS_STOP=2;
     public static final int IS_START=3;
     public static final int IS_PAUSE=4;
-    public static final int EVENT_UPDATE_TIME =10;
-    public static final int EVENT_DELETE_MUSIC=9;
-    public static final int EVENT_FILE_NO_EXISTS=8;
+    public static final int EVENT_UPDATE_TIME =10387;
+    public static final int EVENT_DELETE_MUSIC=937;
+    public static final int EVENT_START_NEW_MUSIC=8156;
+    public static final int EVENT_CHANGE_LOOP_TYPE=11378;
     private int status=NOT_PREPARE;
     private MusicPlayBinder mBinder;
     private MediaPlayer mp;
     private List<Music> playlist;    //播放列表 存放音乐的md5值
-    private Handler handler=null;
+    private List<Handler> handlers;
     private int currentPlayPosition =-1;      //当前播放的位置
     private Music currentMusic;      //当前播放的音乐
     private PlayRunnable currentRunnable;
@@ -61,6 +62,7 @@ public class MusicPlayService extends Service {
         mBinder=new MusicPlayBinder();
         mp=new MediaPlayer();
         playlist=new LinkedList<>();
+        handlers=new ArrayList<>();
 
         mp.setOnPreparedListener(new MediaPlayer.OnPreparedListener() {
             @Override
@@ -72,28 +74,18 @@ public class MusicPlayService extends Service {
     }
 
     private void prepared(){
-        if(currentMusic.getIsServer()==1){
-            boolean login = sp.getBoolean("login",false);
-            if(login){
-                String userId=sp.getString("userId","-1");
-                int musicId=currentMusic.getId();
-                HttpUtil.listenMusic(this,userId,musicId);
-            }
-        }
 
-        //播放歌曲时，结束上一次还没走完的线程
-        if(currentRunnable!=null){
-            currentRunnable.stop();
+        status=IS_PREPARE;
+
+        if(saveComment !=currentPlayPosition){      //如果最后一首和当前要播放的不一致，就播放这首歌
+            playMusic(saveComment);
+            return;
         }
 
         //通知已经准备完毕
-        handler.sendEmptyMessage(IS_PREPARE);
-
-        //将时间设置为00:00
-        Message message=new Message();
-        message.what= EVENT_UPDATE_TIME;
-        message.arg1=0;
-        handler.sendMessage(message);
+        for(Handler handler:handlers){
+            handler.sendEmptyMessage(IS_PREPARE);
+        }
 
         //播放音乐
         status=IS_START;
@@ -119,31 +111,77 @@ public class MusicPlayService extends Service {
         return super.onUnbind(intent);
     }
 
+    private int saveComment = -1;
     private void playMusic(int position) {
-        if(currentPlayPosition==position){
-            mp.seekTo(0);
-            mp.pause();
-            prepared();
+
+        for(Handler handler:handlers){
+            Message msg=new Message();
+            msg.what=EVENT_START_NEW_MUSIC;
+            msg.obj=playlist.get(position);
+            handler.sendMessage(msg);
+        }
+
+        if(status==PREPAREING){          //如果当前还在准备的时候播放下一首，就存起来，不播放，准备完再播放这首歌
+            saveComment =position;           //一直覆盖，播放最后一首
+            return;
+        }
+        saveComment =position;
+
+        if(currentMusic!=null&&currentMusic.getMd5().equals(playlist.get(position).getMd5())){
+            if(status==IS_START||status==IS_PAUSE){
+                mp.seekTo(0);
+                mp.pause();
+                prepared();
+            }
             return;
         }
 
+        //开始准备
+        status=PREPAREING;
         currentPlayPosition =position;
         currentMusic=playlist.get(position);
         try{
             if(currentMusic.getFilePath()!=null){
                 mp.reset();
                 if(currentMusic.getIsServer()==1){
-                    mp.setDataSource(currentMusic.getFilePath());
-                    saveListenCnt();
+
+                    if(currentMusic.getDownloadStatus()==0||currentMusic.getDownloadStatus()==1){
+
+                        File f=new File(currentMusic.getFilePath());
+                        if(f.exists()){
+                            mp.setDataSource(f.getPath());
+                        }else{
+                            Toast.makeText(this, "文件不存在，请重新扫描文件", Toast.LENGTH_SHORT).show();
+                            return;
+                        }
+                    }else if(currentMusic.getDownloadStatus()==3){
+
+                        Music i=localMusicDao.selectMusicByMd5(currentMusic.getMd5());
+                        if(i!=null&&i.getFilePath()!=null&&i.getDownloadStatus()!=2){
+                            File file=new File(i.getFilePath());
+                            if(file.exists()){
+                                mp.setDataSource(file.getPath());
+                            }else{
+                                mp.setDataSource(currentMusic.getFilePath());
+                            }
+                        }else{
+                            mp.setDataSource(currentMusic.getFilePath());
+                        }
+                    }
                 }else{
+
                     File file=new File(currentMusic.getFilePath());
                     if(file.exists()){
+
                         mp.setDataSource(file.getPath());
+
                     }else{
                         Toast.makeText(this, "文件不存在，请重新扫描文件", Toast.LENGTH_SHORT).show();
+                        return;
                     }
                 }
                 mp.prepareAsync();
+
             }
         } catch (IOException e) {
             e.printStackTrace();
@@ -168,10 +206,22 @@ public class MusicPlayService extends Service {
         }
     }
 
-    @Override
-    public int onStartCommand(Intent intent, int flags, int startId) {
-        LogUtil.printLog("onStartCommand");
-        return START_NOT_STICKY;
+    private void playPreviousMusic(){
+        switch (currentLoopType){
+            case LOOP_TYPE_LIST:
+                if(currentPlayPosition-1<0){
+                    playMusic(playlist.size()-1);
+                }else{
+                    playMusic(currentPlayPosition-1);
+                }
+                break;
+            case LOOP_TYPE_RANDOM:
+                playMusic(random.nextInt(playlist.size()));
+                break;
+            case LOOP_TYPE_SINGLE:
+                playMusic(currentPlayPosition);
+                break;
+        }
     }
 
     @Override
@@ -219,21 +269,24 @@ public class MusicPlayService extends Service {
             }
         }
 
+        public void playMusics(List<Music> musics){
+            for(int i=0;i<musics.size();i++){
+                if(i==0){
+                    playMusic(musics.get(i));
+                }else{
+                    nextPlay(musics.get(i));
+                }
+            }
+        }
+
         //下一首播放
         public void nextPlay(Music item){
+
             if(status==NOT_PREPARE){
                 playlist.add(item);
                 MusicPlayService.this.playMusic(0);
             }else{
-
-                boolean f=false;
-                for(int i=0;i<playlist.size();i++){
-                    if(playlist.get(i).getMd5().equals(item.getMd5())){
-                        f=true;
-                        break;
-                    }
-                }
-                if(!f){
+                if(!playlist.contains(item)){
                     playlist.add(currentPlayPosition +1,item);
                 }
             }
@@ -241,34 +294,34 @@ public class MusicPlayService extends Service {
 
         //从播放列表中删除一首音乐
         public void deleteMusic(String md5){
+            boolean exists=false;
+            for(Music item:playlist){
+                if(item.getMd5().equals(md5)){
+                    exists=true;
+                    break;
+                }
+            }
+
+            if(!exists){     //如果存在于播放列表中才删除
+                return;
+            }
+
             if(playlist.size()==1){
-                LogUtil.printLog("删除全部");
                 deleteAllMusic();
                 return;
             }
 
-            if(md5.equals(currentMusic.getMd5())){
-                int index=0;//其实没意义
-                switch (getCurrentLoopType()){
-                    case MusicPlayService.LOOP_TYPE_LIST:
-                    case MusicPlayService.LOOP_TYPE_SINGLE:
-                        index=(getCurrentMusicPosition()+1)%getPlayList().size();
-                        break;
-                    case MusicPlayService.LOOP_TYPE_RANDOM:
-                        while((index=random.nextInt(getPlayList().size()))==getCurrentMusicPosition());
-                        break;
-                }
-                MusicPlayService.this.playMusic(index);
+
+            playlist.remove(currentPlayPosition);
+            if(currentPlayPosition<playlist.size()){
+                MusicPlayService.this.playMusic(currentPlayPosition);
+            }else{
+                MusicPlayService.this.playMusic(currentPlayPosition-1);
             }
 
-            for(Music item:playlist){
-                if(item.getMd5().equals(md5)){
-                    playlist.remove(item);
-                    break;
-                }
+            for(Handler handler:handlers){
+                handler.sendEmptyMessage(EVENT_DELETE_MUSIC);
             }
-            currentPlayPosition=playlist.indexOf(currentMusic);
-            handler.sendEmptyMessage(EVENT_DELETE_MUSIC);
         }
 
         public void deleteAllMusic(){
@@ -282,9 +335,13 @@ public class MusicPlayService extends Service {
             //设置状态
             status=NOT_PREPARE;
 
+            currentMusic=null;
+
             //重置位置
             currentPlayPosition=-1;
-            handler.sendEmptyMessage(EVENT_DELETE_MUSIC);
+            for(Handler handler:handlers){
+                handler.sendEmptyMessage(EVENT_DELETE_MUSIC);
+            }
         }
 
         //刷新
@@ -293,9 +350,21 @@ public class MusicPlayService extends Service {
             //刷新isServer是待确定的项
             for(Music music:playlist){
                 if(music.getIsServer()==2){
-                    music=MusicPlayService.this.localMusicDao.selectMusicByMd5(music.getMd5());
+                    music=localMusicDao.selectMusicByMd5(music.getMd5());
                 }
             }
+        }
+
+        public void seekTo(int t){
+            mp.seekTo(t);
+        }
+
+        public void playNext(){
+            playNextMusic();
+        }
+
+        public void playPrevious(){
+            playPreviousMusic();
         }
 
         public int getStatus(){
@@ -310,12 +379,16 @@ public class MusicPlayService extends Service {
             return currentPlayPosition;
         }
 
-        public void setHandler(Handler handler){
-            MusicPlayService.this.handler=handler;
+        public void addHandler(Handler handler){
+            handlers.add(handler);
+        }
+
+        public void removeHandler(Handler handler){
+            handlers.remove(handler);
         }
 
         public void setLocalMusicDao(LocalMusicDao dao){
-            MusicPlayService.this.localMusicDao=dao;
+            localMusicDao=dao;
         }
 
         public int getCurrentMusicDuration(){
@@ -336,18 +409,21 @@ public class MusicPlayService extends Service {
 
         public void setCurrentLoopType(int loopType){
             MusicPlayService.this.currentLoopType=loopType;
+
+            for(Handler handler:handlers){
+                handler.sendEmptyMessage(EVENT_CHANGE_LOOP_TYPE);
+            }
         }
     }
 
     private class PlayRunnable implements Runnable{
 
-        private boolean f=true;
         @Override
         public void run() {
             while(true){
                 try {
                     //400ms更新一次UI
-                    Thread.sleep(400);
+                    Thread.sleep(250);
 
                     //音乐暂停的时候，时间更新暂停
                     while(status==IS_PAUSE){
@@ -357,16 +433,18 @@ public class MusicPlayService extends Service {
                     if(status!=IS_START)
                         return;
 
-                    if(!f)
+                    if(!this.equals(currentRunnable))
                         return;
 
                     //设置UI
-                    Message message=new Message();
-                    message.what= EVENT_UPDATE_TIME;
-                    message.arg1=mp.getCurrentPosition();
-                    handler.sendMessage(message);
+                    for(Handler handler:handlers){
+                        Message message=new Message();
+                        message.what= EVENT_UPDATE_TIME;
+                        message.arg1=mp.getCurrentPosition();
+                        handler.sendMessage(message);
+                    }
 
-                    if(mp.getCurrentPosition()==mp.getDuration())
+                    if(mp.getCurrentPosition()/1000>=mp.getDuration()/1000)
                         break;
 
                 } catch (InterruptedException e) {
@@ -376,10 +454,7 @@ public class MusicPlayService extends Service {
 
             //自然结束时，根据循环类型选择下一首
             playNextMusic();
-        }
 
-        public void stop(){
-            f=false;
         }
     }
 }
